// src/typechain-types/contracts/EtchMarket.ts
  export type EthscriptionOrderStruct = {
    signer: PromiseOrValue<string>;
    creator: PromiseOrValue<string>;
    ethscriptionId: PromiseOrValue<BytesLike>;
    quantity: PromiseOrValue<BigNumberish>;
    currency: PromiseOrValue<string>;
    price: PromiseOrValue<BigNumberish>;
    nonce: PromiseOrValue<BigNumberish>;
    startTime: PromiseOrValue<BigNumberish>;
    endTime: PromiseOrValue<BigNumberish>;
    protocolFeeDiscounted: PromiseOrValue<BigNumberish>;
    creatorFee: PromiseOrValue<BigNumberish>;
    params: PromiseOrValue<BytesLike>;
    v: PromiseOrValue<BigNumberish>;
    r: PromiseOrValue<BytesLike>;
    s: PromiseOrValue<BytesLike>;
  };

  export type EthscriptionOrderStructOutput = [
    string,
    string,
    string,
    BigNumber,
    string,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    number,
    number,
    string,
    number,
    string,
    string
  ] & {
    signer: string;
    creator: string;
    ethscriptionId: string;
    quantity: BigNumber;
    currency: string;
    price: BigNumber;
    nonce: BigNumber;
    startTime: BigNumber;
    endTime: BigNumber;
    protocolFeeDiscounted: number;
    creatorFee: number;
    params: string;
    v: number;
    r: string;
    s: string;
  };
// ...
export declare namespace BatchOrder {
  export type EthscriptionOrderStruct = {
    signer: PromiseOrValue<string>;
    creator: PromiseOrValue<string>;
    ethscriptionIds: PromiseOrValue<BytesLike>[];
    quantities: PromiseOrValue<BigNumberish>[];
    currency: PromiseOrValue<string>;
    price: PromiseOrValue<BigNumberish>;
    nonce: PromiseOrValue<BigNumberish>;
    startTime: PromiseOrValue<BigNumberish>;
    endTime: PromiseOrValue<BigNumberish>;
    protocolFeeDiscounted: PromiseOrValue<BigNumberish>;
    creatorFee: PromiseOrValue<BigNumberish>;
    params: PromiseOrValue<BytesLike>;
    v: PromiseOrValue<BigNumberish>;
    r: PromiseOrValue<BytesLike>;
    s: PromiseOrValue<BytesLike>;
  };

  export type EthscriptionOrderStructOutput = [
    string,
    string,
    string[],
    BigNumber[],
    string,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    number,
    number,
    string,
    number,
    string,
    string
  ] & {
    signer: string;
    creator: string;
    ethscriptionIds: string[];
    quantities: BigNumber[];
    currency: string;
    price: BigNumber;
    nonce: BigNumber;
    startTime: BigNumber;
    endTime: BigNumber;
    protocolFeeDiscounted: number;
    creatorFee: number;
    params: string;
    v: number;
    r: string;
    s: string;
  };

  export type MerkleTreeNodeStruct = {
    value: PromiseOrValue<BytesLike>;
    position: PromiseOrValue<BigNumberish>;
  };

  export type MerkleTreeNodeStructOutput = [string, number] & {
    value: string;
    position: number;
  };

  export type MerkleTreeStruct = {
    root: PromiseOrValue<BytesLike>;
    proof: BatchOrder.MerkleTreeNodeStruct[];
  };

  export type MerkleTreeStructOutput = [
    string,
    BatchOrder.MerkleTreeNodeStructOutput[]
  ] & { root: string; proof: BatchOrder.MerkleTreeNodeStructOutput[] };
}